pyproject.toml
[project]
name = "llama-inference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.0",
    "datasets>=4.2.0",
    "hf-transfer>=0.1.9",
    "loguru>=0.7.3",
    "pydantic>=2.12.2",
    "pyyaml>=6.0.3",
    "vllm>=0.11.0",
    "bitsandbytes; sys_platform == 'linux'",
    "peft>=0.17.1",
]

[project.scripts]
infer = "llama_inference.cli.inference:main"

[tool.setuptools]
package-dir = {"" = "src"}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
